# SistemaWebLaboratorio/Software/backend/openapi-dialogflow.yaml
openapi: 3.0.0
info:
  title: Laboratorio Clínico Franz - API para Dialogflow CX
  description: |
    API REST para integración con Dialogflow CX.
    Permite consultar precios de exámenes, valores de referencia,
    disponibilidad de horarios y agendar citas.
  version: 1.0.0
  contact:
    name: Soporte Técnico
    email: soporte@laboratoriofranz.com

servers:
  - url: https://api.laboratoriofranz.com
    description: Servidor de producción
  - url: http://localhost:3000
    description: Servidor de desarrollo

tags:
  - name: Exámenes
    description: Operaciones relacionadas con exámenes de laboratorio
  - name: Citas
    description: Operaciones relacionadas con agendamiento de citas

paths:
  /api/dialogflow/examenes:
    get:
      operationId: listarExamenes
      summary: Listar todos los exámenes disponibles
      description: Devuelve la lista de exámenes con sus precios actuales
      tags:
        - Exámenes
      responses:
        '200':
          description: Lista de exámenes obtenida exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListaExamenesResponse'

  /api/dialogflow/examenes/precio/{nombre}:
    get:
      operationId: consultarPrecioExamen
      summary: Consultar precio de un examen
      description: Busca un examen por nombre y devuelve su precio actual
      tags:
        - Exámenes
      parameters:
        - name: nombre
          in: path
          required: true
          description: Nombre del examen (búsqueda parcial)
          schema:
            type: string
          example: hemograma
      responses:
        '200':
          description: Precio del examen encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrecioExamenResponse'

  /api/dialogflow/examenes/rango/{nombre}:
    get:
      operationId: consultarRangoExamen
      summary: Consultar valores de referencia de un examen
      description: Devuelve los rangos normales/valores de referencia de un examen
      tags:
        - Exámenes
      parameters:
        - name: nombre
          in: path
          required: true
          description: Nombre del examen
          schema:
            type: string
          example: glucosa
      responses:
        '200':
          description: Valores de referencia del examen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RangoExamenResponse'

  /api/dialogflow/citas/disponibilidad:
    get:
      operationId: consultarDisponibilidad
      summary: Consultar disponibilidad de horarios
      description: Devuelve los horarios disponibles para una fecha específica
      tags:
        - Citas
      parameters:
        - name: fecha
          in: query
          required: true
          description: Fecha a consultar en formato YYYY-MM-DD
          schema:
            type: string
            format: date
          example: '2025-01-15'
        - name: servicio
          in: query
          required: false
          description: Código del servicio (opcional)
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Disponibilidad consultada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisponibilidadResponse'

  /api/dialogflow/citas/agendar:
    post:
      operationId: agendarCita
      summary: Agendar una nueva cita
      description: Crea una nueva cita para un paciente en la fecha y hora especificadas
      tags:
        - Citas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgendarCitaRequest'
      responses:
        '200':
          description: Cita agendada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CitaResponse'
        '400':
          description: Datos inválidos o no hay disponibilidad
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/dialogflow/citas/paciente/{cedula}:
    get:
      operationId: consultarCitasPaciente
      summary: Consultar citas de un paciente
      description: Devuelve las citas pendientes de un paciente identificado por su cédula
      tags:
        - Citas
      parameters:
        - name: cedula
          in: path
          required: true
          description: Cédula del paciente
          schema:
            type: string
          example: '1234567890'
      responses:
        '200':
          description: Citas del paciente obtenidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListaCitasResponse'

  /api/dialogflow/webhook:
    post:
      operationId: handleDialogflowWebhook
      summary: Webhook para Dialogflow CX
      description: Endpoint que recibe las solicitudes de fulfillment de Dialogflow CX
      tags:
        - Webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DialogflowWebhookRequest'
      responses:
        '200':
          description: Respuesta de fulfillment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DialogflowWebhookResponse'

components:
  schemas:
    # Base Response
    BaseResponse:
      type: object
      required:
        - success
        - mensaje
      properties:
        success:
          type: boolean
          description: Indica si la operación fue exitosa
        mensaje:
          type: string
          description: Mensaje descriptivo para el chatbot
        error_code:
          type: string
          description: Código de error si aplica

    # Precio Examen
    PrecioExamen:
      type: object
      properties:
        nombre:
          type: string
          description: Nombre del examen
          example: Hemograma Completo
        precio:
          type: number
          format: float
          description: Precio actual en USD
          example: 15.00
        moneda:
          type: string
          description: Moneda
          example: USD
        categoria:
          type: string
          description: Categoría del examen
          example: Hematología
        requiere_ayuno:
          type: boolean
          description: Si requiere ayuno
          example: false
        horas_ayuno:
          type: integer
          description: Horas de ayuno requeridas
          example: 8

    PrecioExamenResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PrecioExamen'

    ListaExamenesResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/PrecioExamen'

    # Rango Referencia
    RangoReferencia:
      type: object
      properties:
        nombre:
          type: string
          description: Nombre del examen
          example: Glucosa en Ayunas
        valor_min:
          type: number
          format: float
          description: Valor mínimo de referencia
          example: 70
        valor_max:
          type: number
          format: float
          description: Valor máximo de referencia
          example: 100
        unidad_medida:
          type: string
          description: Unidad de medida
          example: mg/dL
        valores_texto:
          type: string
          description: Descripción textual del rango
          example: Negativo

    RangoExamenResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/RangoReferencia'

    # Disponibilidad
    HorarioDisponible:
      type: object
      required:
        - slot_id
        - hora
        - cupos
        - sede
        - servicio
      properties:
        slot_id:
          type: integer
          description: ID del slot
          example: 123
        hora:
          type: string
          description: Hora de inicio
          example: '09:00'
        cupos:
          type: integer
          description: Cupos disponibles
          example: 3
        sede:
          type: string
          description: Nombre de la sede
          example: Sede Principal
        servicio:
          type: string
          description: Nombre del servicio
          example: Toma de muestras

    Disponibilidad:
      type: object
      properties:
        fecha:
          type: string
          format: date
          description: Fecha consultada
          example: '2025-01-15'
        total_horarios:
          type: integer
          description: Total de horarios disponibles
          example: 8
        horarios:
          type: array
          items:
            $ref: '#/components/schemas/HorarioDisponible'

    DisponibilidadResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Disponibilidad'

    # Citas
    AgendarCitaRequest:
      type: object
      required:
        - fecha
        - hora
        - cedula_paciente
      properties:
        fecha:
          type: string
          format: date
          description: Fecha de la cita en formato YYYY-MM-DD
          example: '2025-01-15'
        hora:
          type: string
          description: Hora de la cita en formato HH:MM
          example: '09:00'
        cedula_paciente:
          type: string
          description: Cédula del paciente
          example: '1234567890'
        examen:
          type: string
          description: Nombre del examen o servicio
          example: Hemograma Completo
        codigo_servicio:
          type: integer
          description: Código del servicio
          example: 1
        codigo_sede:
          type: integer
          description: Código de la sede
          example: 1
        observaciones:
          type: string
          description: Observaciones adicionales

    CitaInfo:
      type: object
      properties:
        codigo_cita:
          type: integer
          description: Código único de la cita
          example: 456
        fecha:
          type: string
          format: date
          description: Fecha de la cita
          example: '2025-01-15'
        hora:
          type: string
          description: Hora de la cita
          example: '09:00'
        servicio:
          type: string
          description: Servicio
          example: Toma de muestras
        sede:
          type: string
          description: Sede
          example: Sede Principal
        estado:
          type: string
          description: Estado de la cita
          example: AGENDADA
          enum:
            - AGENDADA
            - CONFIRMADA
            - EN_PROCESO
            - COMPLETADA
            - CANCELADA

    CitaResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/CitaInfo'

    ListaCitasResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/CitaInfo'

    # Error
    ErrorResponse:
      type: object
      required:
        - success
        - mensaje
        - error_code
      properties:
        success:
          type: boolean
          example: false
        mensaje:
          type: string
          description: Mensaje de error descriptivo
        error_code:
          type: string
          description: Código de error
          enum:
            - EXAMEN_NO_ENCONTRADO
            - PACIENTE_NO_ENCONTRADO
            - FECHA_INVALIDA
            - FECHA_PASADA
            - SLOT_NO_DISPONIBLE
            - CITA_DUPLICADA
            - ERROR_INTERNO

    # Dialogflow CX Webhook
    DialogflowWebhookRequest:
      type: object
      description: Solicitud de fulfillment de Dialogflow CX
      properties:
        fulfillmentInfo:
          type: object
          properties:
            tag:
              type: string
              description: Tag del fulfillment
              example: consultar-precio
        sessionInfo:
          type: object
          properties:
            parameters:
              type: object
              additionalProperties: true
              description: Parámetros extraídos por Dialogflow

    DialogflowWebhookResponse:
      type: object
      description: Respuesta de fulfillment para Dialogflow CX
      properties:
        fulfillment_response:
          type: object
          properties:
            messages:
              type: array
              items:
                type: object
                properties:
                  text:
                    type: object
                    properties:
                      text:
                        type: array
                        items:
                          type: string
        sessionInfo:
          type: object
          properties:
            parameters:
              type: object
              additionalProperties: true